requirements.txt
openai
crewai
langchain
python-dotenv
langchain_openai